Update: Minor changes. Improve design the overall layout.
Update the dependencies to remove unused dependencies. Add new
dependencies based on what is or might be needed. Update the lock
file accordingly.

Define base error, and a child class to contain errors - if any
occurs with the API. Add tests to test these changes so far.

Modify the pre-commit hook to generate dev-requirements before
every commit. Generation of the test report depends upon the
test cases compiling (which in turn needs all the developer
packages to run the tests).

Add a configuration file for Codecov, will prevent from stupidly
failing tests if the coverage drops by a bit - seriously, it
expects me to maintain a 100% coverage, or it fails the test.

Welp, no more stupidity.