Update: Minor refactor. Define user avatar. Modify and add tests
Minor refactor, includes reducing the line length to fit at 88
characters. Also create a new package to hold common class definitions
together.

Define the base object class, this will be the base class for almost
all objects. Abstract class as its only purpose is to act as a parent
class and provide common functionality.

Define a class for user avatar, will be used to contain data regarding
profile pictures in case of users. Add new test cases to fit in with
the modifications being made to the code.