Update: Add new types to map API objects. Minor bug patches
Define a new base class for abstract enum's. Will be used to
define enums that are somewhat similar to `BaseObject` in terms
of API and public method(s). Add comments for the corresponding
changes made to the code.

Define new classes to map fuzzy date, media status, sort, format,
source and season. All the media type classes are internally based
on abstract enum defined in this commit.

Add test cases to test changes made so far.