Update: Modifications to makefile. Map new types. Add tests. BugFix
Add a new field `mypy` to makefile - simply runs mypy with the
default (makefile) configurations on the entire project.

Create a default logger at the root of the project.

Separate `BaseEnum` and `BaseObject` to deal with them separately.
Add corresponding docstrings to better document code.

Patch a bug in `BaseObject`, instead of directly attemping to JSON
decode every instance variable present in the object, based on the
type of the variable present in the class instance, different
strategies will be used to be able to map it into a printable
string.

Add new method to `BaseEnum` to be able to map a string to an enum
entry, this method will iterate over all entries present in a class
derived from BaseEnum, and compare the `key` with the translation of
every entry present in the enum. As soon as a match is found, the
enum entry will be returned. If no match is present, an error will
be thrown.

Patch some bugs in the common modules being used for running tests.
Includes type-casting a deque into a list, and then unparsing the
string before using it as a method argument.

Add new enums and classes to map the API response parameters into
objects. Add new test-cases.

The class `MediaData` is a shell object currently, a few objects
are yet to be mapped before the `MediaData` class can be fully
created.