BugPatch: Patch an internal bug. Add test cases.
Patch an internal bug in `stringify` method of base object. The bug
occurred becuase of poor type-matching being used to compare the
type of objects while attempting to map an instance of a class into
string.

Patch this by a slight change in the logic of the method, along with
improved documentation to ease any changes made to the module in the
future.

Add test cases and a generic method for tests to ensure better
testing and reliability of the codebase.